add fcm

+++ application/views/administrator/mod_alumni/Data.php
@@ -51,6 +51,7 @@
     								<th>Telepon</th>
     								<th>Tahun Mondok</th>
     								<th>Tahun Keluar</th>
+                                    <th>Actions</th>
     							</tr>
     						</thead>
     						<tbody>
@@ -71,7 +72,7 @@
     											
     											<button type="button" class="btn btn-default" disabled=""><i class="fa fa-align-center"></i></button>
 
-    											<a href="<?=base_url()?>administrator/hapus_alumni/<?=$a['id_alumni']?>" class="btn btn-danger">Hapus</a>
+    											<a href="<?=base_url()?>administrator/hapus_alumni/<?=$a['id_alumni']?>" class="btn btn-danger" onclick="return confirm('Anda Yakin Ingin Menghapus Data Ini ?')">Hapus</a>
     										</div>
     									</td>
     								</tr>
