+++ application/views/administrator/mod_lembaga_nj/Data.php
@@ -64,7 +64,7 @@
     											
     											<button type="button" class="btn btn-default" disabled=""><i class="fa fa-align-center"></i></button>
 
-    											<a href="<?=base_url()?>administrator/hapus_lembaga_nj/<?=$l['id_lembaga']?>" class="btn btn-danger">Hapus</a>
+    											<a href="<?=base_url()?>administrator/hapus_lembaga_nj/<?=$l['id_lembaga']?>" class="btn btn-danger" onclick="return confirm('Anda Yakin Ingin Menghapus Data Ini ?')">Hapus</a>
     										</div>
     									</td>
     								</tr>
