+++ application/views/administrator/mod_visi_misi/Data.php
@@ -48,7 +48,7 @@
     											
     											<button type="button" class="btn btn-default" disabled=""><i class="fa fa-align-center"></i></button>
 
-    											<a href="<?=base_url()?>administrator/hapus_visi_misi/<?=$l['id_visi_misi']?>?lembaga=<?=$this->session->userdata('username')?>" class="btn btn-danger">Hapus</a>
+    											<a href="<?=base_url()?>administrator/hapus_visi_misi/<?=$l['id_visi_misi']?>?lembaga=<?=$this->session->userdata('username')?>" class="btn btn-danger" onclick="return confirm('Anda Yakin Ingin Menghapus Data Ini ?')">Hapus</a>
     										</div>
     									</td>
     								</tr>
